+++ SDLProject/src/Player.cpp
@@ -37,7 +37,7 @@ void Player::Draw()
 	Player_SpriteAnimation->Draw(GameObject_Transform->getX(), GameObject_Transform->getY(), GameObject_Width, GameObject_Height, GameObject_Flip, 1.0, 1.0);
 }
 
-void Player::Update(double dt)
+void Player::Update(Uint64 dt)
 {
 	Player_IsWalking = false;
 	Player_RigidBody->SetForceToZero();
